refactor(lib): Move a few source files into lib/
Both bin/ and config/ directories have been left alone.
