+++ bin/server.js
@@ -5,8 +5,8 @@
 var config = require('../config')
 var dbServer = require('fxa-auth-db-server')
 var error = dbServer.errors
-var logger = require('../logging')('bin.server')
-var DB = require('../db')(logger, error)
+var logger = require('../lib/logging')('bin.server')
+var DB = require('../lib/db')(logger, error)
 
 function shutdown() {
   process.nextTick(process.exit)
